backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Caching==2.1.0
SQLAlchemy==2.0.23
pymysql==1.1.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.0.0
Werkzeug==3.0.1
cryptography>=41.0.0
python-dotenv>=1.0.0

# 测试依赖
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.2.0

